% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TechanJS.R
\name{renderTechanJS}
\alias{renderTechanJS}
\title{Widget render function for use in Shiny}
\usage{
renderTechanJS(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

